finally have storage proto working
