+++ src/service/src/grpc.rs
@@ -9,21 +9,18 @@
 
 //! gRPC transport for the [client](crate::client) module.
 
-use std::fmt::{self, Debug};
-use std::pin::Pin;
-use std::sync::Arc;
-use std::time::Duration;
-
 use async_stream::stream;
 use async_trait::async_trait;
 use futures::future;
 use futures::stream::{Stream, StreamExt, TryStreamExt};
 use http::uri::PathAndQuery;
 use mz_ore::netio::{Listener, SocketAddr, SocketAddrType};
-use mz_proto::{ProtoType, RustType, TryFromProtoError};
+use mz_proto::{ProtoType, RustType};
 use once_cell::sync::Lazy;
 use semver::Version;
-use serde::{Deserialize, Serialize};
+use std::fmt::{self, Debug};
+use std::pin::Pin;
+use std::sync::Arc;
 use tokio::net::UnixStream;
 use tokio::select;
 use tokio::sync::mpsc::{self, UnboundedSender};
@@ -40,6 +37,7 @@ use tracing::{debug, error, info};
 
 use crate::client::{GenericClient, Partitionable, Partitioned};
 use crate::codec::{StatCodec, StatsCollector};
+use crate::params::GrpcClientParameters;
 
 include!(concat!(env!("OUT_DIR"), "/mz_service.params.rs"));
 
@@ -420,62 +418,5 @@ impl Interceptor for VersionCheckExactInterceptor {
                 version, self.version
             ))),
         }
-    }
-}
-
-/// gRPC client parameters.
-#[derive(Serialize, Deserialize, Clone, Debug, Default, PartialEq, Eq)]
-pub struct GrpcClientParameters {
-    /// Timeout to apply to initial connection establishment.
-    pub connect_timeout: Option<Duration>,
-    /// Time waited after the last received HTTP/2 frame before sending
-    /// a keep-alive PING to the server. Note that this is an HTTP/2
-    /// keep-alive, and is separate from TCP keep-alives.
-    pub http2_keep_alive_interval: Option<Duration>,
-    /// Time waited without response after a keep-alive PING before
-    /// terminating the connection.
-    pub http2_keep_alive_timeout: Option<Duration>,
-}
-
-impl GrpcClientParameters {
-    pub fn update(&mut self, other: Self) {
-        let Self {
-            connect_timeout,
-            http2_keep_alive_interval,
-            http2_keep_alive_timeout,
-        } = self;
-        let Self {
-            connect_timeout: other_connect_timeout,
-            http2_keep_alive_interval: other_http2_keep_alive_interval,
-            http2_keep_alive_timeout: other_http2_keep_alive_timeout,
-        } = other;
-
-        if let Some(v) = other_connect_timeout {
-            *connect_timeout = Some(v);
-        }
-        if let Some(v) = other_http2_keep_alive_interval {
-            *http2_keep_alive_interval = Some(v);
-        }
-        if let Some(v) = other_http2_keep_alive_timeout {
-            *http2_keep_alive_timeout = Some(v);
-        }
-    }
-}
-
-impl RustType<ProtoGrpcClientParameters> for GrpcClientParameters {
-    fn into_proto(&self) -> ProtoGrpcClientParameters {
-        ProtoGrpcClientParameters {
-            connect_timeout: self.connect_timeout.into_proto(),
-            http2_keep_alive_interval: self.http2_keep_alive_interval.into_proto(),
-            http2_keep_alive_timeout: self.http2_keep_alive_timeout.into_proto(),
-        }
-    }
-
-    fn from_proto(proto: ProtoGrpcClientParameters) -> Result<Self, TryFromProtoError> {
-        Ok(Self {
-            connect_timeout: proto.connect_timeout.into_rust()?,
-            http2_keep_alive_interval: proto.http2_keep_alive_interval.into_rust()?,
-            http2_keep_alive_timeout: proto.http2_keep_alive_timeout.into_rust()?,
-        })
     }
 }
