+++ src/service/src/lib.rs
@@ -84,6 +84,7 @@ pub mod client;
 pub mod codec;
 pub mod grpc;
 pub mod local;
+pub mod params;
 pub mod retry;
 pub mod secrets;
 pub mod tracing;
